websockets/constructor/007.html

<!doctype html>
<title>WebSockets: new WebSocket(url, null char)</title>
<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>
<script src=../constants.js></script>
<div id=log></div>
<script>
var domain = SCHEME_DOMAIN_PORT.split('://')[1];
test(function() {assert_throws("SyntaxError", function(){new WebSocket("ws://"+domain+"/", 'a'+String.fromCharCode(0)+'b')})});
</script>
